+++ src/transaction.rs
@@ -1,12 +1,12 @@
 pub struct Transaction {
-    pub period: u32,
+    pub period: u64,
     pub amount: f64,
 }
 
 impl Transaction {
     // A public constructor method
     #[warn(dead_code)]
-    pub fn new(period: u32, amount: f64) -> Transaction {
+    pub fn new(period: u64, amount: f64) -> Transaction {
         Transaction {
             period,
             amount
